show share utilization also on ShareShowModal
