+++ plugins/shared_filesystem_storage/app/assets/stylesheets/shared_filesystem_storage/_application.scss
@@ -60,3 +60,8 @@ td.has-share-utilization {
     }
   }
 }
+
+//rules for ShareShowModal
+table.share-details div.progress {
+  margin-bottom: 3px;
+}
