+++ plugins/shared_filesystem_storage/app/javascript/app/components/shares/item.jsx
@@ -3,6 +3,7 @@ import { policy } from 'policy';
 import { OverlayTrigger, Tooltip } from 'react-bootstrap';
 import * as constants from '../../constants';
 import ShareActions from './actions';
+import ShareUtilization from './utilization';
 
 class RuleTooltip extends React.Component {
   render() {
@@ -23,19 +24,6 @@ class RuleTooltip extends React.Component {
     );
   }
 }
-
-//TODO: use Unit class from plugins/resources
-const byteUnits = [ 'B', 'KiB', 'MiB', 'GiB', 'TiB', 'PiB', 'EiB' ];
-const displayBytes = (value) => {
-  let index = 0;
-  while (value > 1024) {
-    value = value / 1024;
-    index++;
-  }
-  value = Math.round(value * 100) / 100;
-  return `${value} ${byteUnits[index]}`;
-};
-
 
 export default class ShareItem extends React.Component {
   constructor(props){
@@ -77,54 +65,6 @@ export default class ShareItem extends React.Component {
     return constants.isShareStatusPending(props.share.status);
   }
 
-  renderUtilization() {
-    const { data, isFetching, wasRequested } = this.props.utilization;
-    if (!wasRequested) {
-      return;
-    }
-    if (isFetching) {
-      return <div className='progress'>
-        <div className='progress-bar progress-bar-empty'>
-          <span className='spinner' /> Loading...
-        </div>
-      </div>;
-    }
-    if (data == null) {
-      return <div className='progress'>
-        <div className='progress-bar progress-bar-empty'>
-          Unknown
-        </div>
-      </div>;
-    }
-
-    const snapReserveBytes = data.size_reserved_by_snapshots || 0;
-    const shareSizeBytes   = (data.size_total || 0) + snapReserveBytes;
-
-    const shareUsedBytes = data.size_used || 0;
-    const shareUsedPerc  = 100 * (shareUsedBytes / shareSizeBytes);
-    const shareTooltip   = `${displayBytes(shareUsedBytes)} used by files`;
-
-    const snapUsedBytes    = data.size_used_by_snapshots || 0;
-    const snapDisplayBytes = Math.max(snapUsedBytes, snapReserveBytes);
-    const snapDisplayPerc  = 100 * (snapDisplayBytes / shareSizeBytes);
-    const snapTooltip      = (snapUsedBytes > snapReserveBytes)
-      ? `${displayBytes(snapDisplayBytes)} used by snapshots`
-      : `${displayBytes(snapDisplayBytes)} reserved for snapshots`;
-
-    const tooltip = <Tooltip id={`utilizationTooltip-${this.props.share.id}`}>
-      <nobr>{shareTooltip}</nobr><br/><nobr>{snapTooltip}</nobr>
-    </Tooltip>;
-
-    return (
-      <OverlayTrigger overlay={tooltip} placement='top' delayShow={300} delayHide={150}>
-        <div className='progress'>
-          <div className='progress-bar' style={{width: shareUsedPerc + '%'}} />
-          <div className='progress-bar progress-bar-info' style={{width: snapDisplayPerc + '%'}} />
-        </div>
-      </OverlayTrigger>
-    );
-  }
-
   render(){
     let { share, shareNetwork, shareRules, handleDelete, handleForceDelete } = this.props
 
@@ -136,7 +76,7 @@ export default class ShareItem extends React.Component {
         <td>{share.availability_zone}</td>
         <td>{share.share_proto}</td>
         <td className='has-share-utilization'>
-          {this.renderUtilization()}
+          <ShareUtilization utilization={this.props.utilization} shareID={share.id} compact={true} />
           {share.size || 0} GiB
         </td>
         <td>
